feat: dashboard perf improvements
