+++ frontend/src/pages/DashboardsListPage/index.tsx
@@ -1,0 +1,3 @@
+import DashboardsListPage from './DashboardsListPage';
+
+export default DashboardsListPage;
